Reading some API data to display members
