+++ .idea/misc.xml
@@ -29,10 +29,8 @@
         <entry key="../../../../../../layout/compose-model-1657279858159.xml" value="0.20185810810810811" />
         <entry key="../../../../../../layout/compose-model-1657280011649.xml" value="0.9372093023255814" />
         <entry key="../../../../../../layout/compose-model-1657280020166.xml" value="0.20168776371308017" />
-        <entry key="../../../../layout/compose-model-1662148087456.xml" value="0.15244932432432431" />
-        <entry key="../../../../layout/compose-model-1662148760106.xml" value="0.23425925925925925" />
-        <entry key="../../../../layout/compose-model-1662157228918.xml" value="0.22170608108108109" />
-        <entry key="../../../../layout/compose-model-1662157240357.xml" value="0.22170608108108109" />
+        <entry key="../../../../layout/compose-model-1662193572066.xml" value="0.27037037037037037" />
+        <entry key="../../../../layout/compose-model-1662194829880.xml" value="0.27037037037037037" />
         <entry key="app/src/main/java/rs/appsterdam/appsterdam/MainActivity.kt" value="0.28354430379746837" />
       </map>
     </option>
@@ -40,7 +38,7 @@
   <component name="FrameworkDetectionExcludesConfiguration">
     <file type="web" url="file://$PROJECT_DIR$" />
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="jbr-11" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_17_PREVIEW" project-jdk-name="jbr-11" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
